Fix Plane<Real>::ZX

+++ libs/mathlib/include/impl/plane.h
@@ -10,7 +10,7 @@ template <class Real>
 const Plane<Real> Plane<Real>::YZ = Plane<Real>(Vec3<Real>(Real(1), Real(0), Real(0)), Real(0));
 
 template <class Real>
-const Plane<Real> Plane<Real>::ZX = Plane<Real>(Vec3<Real>(Real(0), Real(1), Real(1)), Real(0));
+const Plane<Real> Plane<Real>::ZX = Plane<Real>(Vec3<Real>(Real(0), Real(1), Real(0)), Real(0));
 
 
 template <class Real>
